BitcoinExpeirmentTest file which compare the temporal motifs of the simulation created by the BitcoinGraphDynamics file and the real data sample computed by ParallelTemporaltest file

+++ graph_dynamics/networks/temporalmotif.py
@@ -1,7 +1,7 @@
 import networkx as nx
 import subprocess
 import numpy as np
-
+import os
 class TemporalMotif():
 
 
@@ -16,12 +16,15 @@ class TemporalMotif():
         self.input_motif = self.graphdynamics_to_temporalmotif(GraphDynamics.get_networkx(), input)
         self.output_motif = self.input_motif.replace(".temporalmotif","") + ".temporalmotifcount"
 
-        args1 = "-i:" + self.input_motif
-        args2 = "-o:" + self.output_motif
-        args3 = "-delta:" + str(delta)
 
-        # calling command of snap in c++
-        subprocess.call([self.exe_directory, args1, args2, args3])
+        if os.path.isfile(self.output_motif ) == False:
+
+            args1 = "-i:" + self.input_motif
+            args2 = "-o:" + self.output_motif
+            args3 = "-delta:" + str(delta)
+
+            # calling command of snap in c++
+            subprocess.call([self.exe_directory, args1, args2, args3])
 
 
     def get_motifdata(self):
@@ -36,19 +39,26 @@ class TemporalMotif():
 
         print "Converting to temporal #: " + input
 
-        output_path = input.replace(".gd","") + ".temporalmotif"
+        output_path = input.replace(".gd", "") + ".temporalmotif"
 
-        set_nodes = set(graph.nodes())
-        # indexes = range(0, len(set_nodes))
-        # all_nodes = dict(zip(set_nodes, indexes))
+        #if file dont exist convert it
+        if os.path.isfile(output_path) == False:
 
-        output_file = open(output_path, "w")
-        edges = graph.edges(data=True)
-        edges = sorted(edges, key=self.getKey)
-        for idy, edge in enumerate(edges):
-            time = edge[2]['time']
-            output_file.write(str(edge[0]) + " " + str(edge[1]) + " " + str(time) + "\n")
+            output_file = open(output_path, "w")
 
-        output_file.close()
+            set_nodes = set(graph.nodes())
+            indexes = range(0, len(set_nodes))
+            all_nodes = dict(zip(set_nodes, indexes))
+
+            edges = graph.edges(data=True)
+            edges = sorted(edges, key=self.getKey)
+
+            for idy, edge in enumerate(edges):
+                toa = all_nodes.get(edge[0])
+                froma = all_nodes.get(edge[1])
+                time = edge[2]['time']
+                output_file.write(str(toa) + " " + str(froma) + " " + str(time) + "\n")
+
+            output_file.close()
         print "Temporal File: " + output_path
         return output_path
